feat: Moves tests to be next to components
