--- frontend/src/core/store/rootStore.ts
@@ -1,14 +1,0 @@
-import { defineStore } from 'pinia'
-
-export const useRootStore = defineStore('rootStore', {
-  state: () => {
-    return {
-      name: ''
-    }
-  },
-  actions: {
-    setName(name: string) {
-      this.name = name
-    }
-  }
-})
