+++ frontend/src/modules/client/components/modals/create-client-form-sections/ContactInfo.vue
@@ -69,7 +69,7 @@
 import { reactive } from 'vue'
 import { required, email } from '@vuelidate/validators'
 import useVuelidate from '@vuelidate/core'
-import InputComponent from '@/modules/shared/InputComponent.vue'
+import InputComponent from '@/modules/shared/inputComponent/InputComponent.vue'
 import { FormResponse } from '@/types/form/formResponse'
 import { ClientContactInfo } from '@/types/client/newClient'
