+++ frontend/src/modules/login/components/companyDetails/CompanyDetails.vue
