+++ frontend/src/services/http.ts
@@ -1,4 +1,4 @@
-import { useAuthStore } from '@/core/store/authStore'
+import { useAuthStore } from '@/core/store/authStore/authStore'
 import axios, { AxiosRequestConfig } from 'axios'
 
 interface Response {
