--- frontend/tests/helpers/createWrapper.ts
@@ -1,7 +1,0 @@
-import { shallowMount } from '@vue/test-utils'
-
-export function createNewWrapper(component: any, options = {}) {
-  return shallowMount(component, {
-    ...options
-  })
-}
