+++ frontend/tsconfig.json
@@ -19,5 +19,5 @@
   "vueCompilerOptions": {
     "experimentalExposeScriptSetupContext": true
   },
-  "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue", "tests/**/*.spec.ts"]
+  "include": ["src/**/*.ts", "src/**/*.d.ts", "src/**/*.tsx", "src/**/*.vue", "src/tests/**/*.spec.ts"]
 }
